Final draft

+++ src/main/java/com/goldrental/dto/WalletResponse.java
@@ -4,21 +4,19 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 
 @Getter
 @Setter
 public class WalletResponse {
 
-    int balance;
-    int blocked;
-    int available;
+    BigDecimal balance;
+    BigDecimal blocked;
+    BigDecimal available;
 
     public WalletResponse() {}
-    public WalletResponse(int balance, int blocked, int available) {
+    public WalletResponse(BigDecimal balance, BigDecimal blocked, BigDecimal available) {
         this.balance = balance;
         this.blocked = blocked;
         this.available = available;
-
     }
 }
